requirements.txt
# requirements.txt

requests
pymongo
python-dotenv
apscheduler
beautifulsoup4